embed_flipbook

embed_flipbook

embed_flipbook(chunk_name, break_type = "auto",
  code_file_name = paste0("embedded_flipbooks/", chunk_name, ".R"),
  rmd_path = paste0("embedded_flipbooks/", chunk_name, "_embed.Rmd"),
  title = stringr::str_replace_all(chunk_name, "_|\\.", " "),
  subtitle = "", author = "", url = paste0("embedded_flipbooks/",
  chunk_name, "_embed.html"), height = 325, font_size = 120,
  title_page = F, ...)

Arguments

chunk_name
a character string referring named chunk containing code to 'flipbook'

break_type
"auto" is default finding appropriate breakpoints, "user" can be used with the special comment message #BREAK within the source code chunk, "non_seq" can be used for non sequential display of code with special comment messages #BREAK2 (will show in second frame) and #BREAK3 (will show in third frame), an integer input can be given too, to simply display the source code chunk multiple times which is appropriate for observing multiple realizations of sampling, "rotate" allows cycling through different lines of code, the comment #ROTATE is used for lines to by cycled through

code_file_name
a .R file path where chunk's code will be saved

rmd_path
an .Rmd path were source of mini flipbook will be saved

title
a character string if a title is desired for the embedded flipbook, defaults to modified chunk name

subtitle
a character string for the embedded flipbook's subtitle, defaults to ""

author
a character string for the embedded flipbook's author info, defaults to ""

url
path to .html rendered mini flipbook

height
numeric size of iframe, defaults to 325

font_size
numeric to adjust the size of code in embedded flipbooks

title_page
logical indicating whether to include a title page for the mini flipbook, defaults to FALSE

...
inherits from chunk_reveal()